feat/Gacha
Added files:
  - gacha.py

Added folder:
  - servers
  - Note that the servers underneath are the server ID's, these are servers created by the bot and therefore, not counted as an addition by a member of the team

Details:

Started working on the gacha system. The folder servers stores the serverIDs of each server to keep track of the users using the gacha system as well as the usersID ( these userID's are .txt files). The .txt files contains names of the characters they have collected )